+++ model/modules/__init__.py
@@ -1,9 +1,13 @@
 from .audio_encoder import AudioEncoder, AudioFrontEnd
 from .audio_generator import AudioGenerator
 from .diffusion import DiffusionModule, NoiseScheduler, UNet
+from .dit import DiT, DiTBlock, PatchEmbed2D, TimestepEmbedder, sincos_pos_embed_2d
+from .flow import ODESampler, RectifiedFlow
+from .latent_diffusion import LatentImageGenerator, build_image_generator
 from .moe import MoEFeedForward, build_ffn
 from .multimodal_projector import MultimodalProjector
 from .transformer import CausalSelfAttention, FeedForward, RMSNorm, TransformerBlock
+from .vae import KLVAE, DiagonalGaussian
 from .video_generator import TemporalAttention, VideoGenerator
 from .vision_encoder import PatchEmbedding, VisionEncoder
 
@@ -19,6 +23,17 @@ __all__ = [
     "DiffusionModule",
     "UNet",
     "NoiseScheduler",
+    "KLVAE",
+    "DiagonalGaussian",
+    "DiT",
+    "DiTBlock",
+    "PatchEmbed2D",
+    "TimestepEmbedder",
+    "sincos_pos_embed_2d",
+    "RectifiedFlow",
+    "ODESampler",
+    "LatentImageGenerator",
+    "build_image_generator",
     "VideoGenerator",
     "TemporalAttention",
     "AudioEncoder",